requirements.txt
cassandra-driver==3.29.0
confluent-kafka==2.3.0
dash==2.15.0
fastapi[all]==0.109.2
fastavro==1.9.3
httpie==3.2.2
kafka-python==2.0.2
pandas==2.2.0
plotly==5.18.0
python-dotenv==1.0.1
numpy==1.26.4
waitress==3.0.0
